+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveServiceImpl.java
@@ -442,6 +442,7 @@ public class StockMoveServiceImpl implements StockMoveService {
     }
 
     String newStockSeq = null;
+    stockMoveLineService.checkTrackingNumber(stockMove);
     stockMoveLineService.checkConformitySelection(stockMove);
     stockMoveLineService.checkExpirationDates(stockMove);